refactor(FE): Container 내부에 Scroll 컴포넌트로 분리 (#146)
