+++ client/src/components/Modal/ReviewModal/ReviewContainer/ReviewContainer.tsx
@@ -1,28 +1,14 @@
-import React, { useMemo } from "react";
-import useReviewInfiniteScroll from "@/hooks/useReviewInfiniteScroll";
-import { ReviewInfo, ReviewScroll } from "@/types/review";
-import PostLoader from "@/components/PostBar/PostLoader/PostLoader";
+import React from "react";
+import ReviewScroll from "@/components/Modal/ReviewModal/ReviewContainer/ReviewScroll/ReviewScroll";
 
 interface ReviewContainerProps {
   postId: string;
 }
 
 const ReviewContainer = ({ postId }: ReviewContainerProps): JSX.Element => {
-  const { data, onIntersect } = useReviewInfiniteScroll(postId);
-  const reviewInfos = useMemo(
-    (): ReviewInfo[] =>
-      data?.pages.flatMap(
-        (reviewScroll: ReviewScroll) => reviewScroll.reviews
-      ) ?? [],
-    [data]
-  );
-
   return (
     <div className="review-modal__review-container">
-      {reviewInfos.map((reviewInfo: ReviewInfo) => (
-        <div key={reviewInfo.id}>hi</div>
-      ))}
-      <PostLoader onIntersect={onIntersect} />
+      <ReviewScroll postId={postId} />
     </div>
   );
 };
